Refactor command output logging
Previously, when chef command has been run, no logging took place
while the command was running. Command output was reported in bulk at
the end of the operation.

This commit removes the bulk report operation with incremental logging
of output, which should feel more natural to the users, since output
will be presented more or less in the real time.